chore: use http::uri::Authority
